--- opencog/reasoning/CMakeLists.txt
@@ -1,7 +1,0 @@
-IF (HAVE_GUILE)
-	ADD_SUBDIRECTORY (engine)
-	ADD_SUBDIRECTORY (pln)
-	ADD_SUBDIRECTORY (r2l)
-ENDIF (HAVE_GUILE)
-
-
