--- opencog/reasoning/README.md
@@ -1,13 +1,0 @@
-OpenCog Reasoning README
-========================
-
-In-progress implementation of a general unified rule engine, to be
-utilized for PLN, Relex2Logic, and RelEx. For more details, view the
-README in the engine subdirectory, and also refer to
-http://wiki.opencog.org/w/Unified_rule_engine
-
-This directory contains the following subdirectories:
-
-* engine -- The core Rule Engine code
-* pln -- The PLN specific code
-* r2l -- The Relex-to-Logic specific
