+++ opencog/unified-rule-engine/Rule.cc
