+++ opencog/unified-rule-engine/Rule.h
