+++ opencog/unified-rule-engine/URECommons.cc
