+++ opencog/unified-rule-engine/URECommons.h
