+++ opencog/unified-rule-engine/backwardchainer/BCPatternMatch.cc
