+++ opencog/unified-rule-engine/backwardchainer/BCPatternMatch.h
