+++ opencog/unified-rule-engine/backwardchainer/BackwardChainer.h
@@ -23,7 +23,7 @@
 #ifndef BACKWARDCHAINER_H_
 #define BACKWARDCHAINER_H_
 
-#include <opencog/reasoning/engine/Rule.h>
+#include <opencog/unified-rule-engine/Rule.h>
 
 class BackwardChainerUTest;
