+++ opencog/unified-rule-engine/forwardchainer/DefaultForwardChainerCB.cc
