+++ opencog/unified-rule-engine/forwardchainer/FCMemory.cc
