+++ opencog/unified-rule-engine/forwardchainer/ForwardChainInputMatchCB.cc
