+++ opencog/unified-rule-engine/forwardchainer/ForwardChainInputMatchCB.h
