+++ opencog/unified-rule-engine/forwardchainer/ForwardChainPatternMatchCB.cc
