+++ opencog/unified-rule-engine/forwardchainer/ForwardChainPatternMatchCB.h
