+++ opencog/unified-rule-engine/forwardchainer/ForwardChainerCallBack.h
