--- tests/reasoning/CMakeLists.txt
@@ -1,2 +1,0 @@
-ADD_SUBDIRECTORY(engine)
-
