+++ tests/unified-rule-engine/CMakeLists.txt
@@ -14,5 +14,5 @@ IF (HAVE_GUILE)
 	ADD_CXXTEST(FCMemoryUTest)
 ENDIF (HAVE_GUILE)
 
-CONFIGURE_FILE(${CMAKE_SOURCE_DIR}/tests/reasoning/engine/test_cpolicy.json
-               ${PROJECT_BINARY_DIR}/tests/reasoning/engine/test_cpolicy.json)
+CONFIGURE_FILE(${CMAKE_SOURCE_DIR}/tests/unified-rule-engine/test_cpolicy.json
+               ${PROJECT_BINARY_DIR}/tests/unified-rule-engine/test_cpolicy.json)
