+++ tests/unified-rule-engine/URECommonsUTest.cxxtest
@@ -5,7 +5,7 @@
  *      Author: misgana
  */
 
-#include <opencog/reasoning/engine/URECommons.h>
+#include <opencog/unified-rule-engine/URECommons.h>
 #include <opencog/guile/SchemeEval.h>
 #include <opencog/guile/SchemeSmob.h>
 #include <opencog/atomspace/AtomSpace.h>
